# Avoid alloc on creating hashable func