--- a/aggregators/models.go
+++ b/aggregators/models.go
@@ -266,7 +266,7 @@ type TransactionAggregationKey struct {
 	TransactionType   string
 	TransactionResult string
 
-	FAASColdstart   nullable.NullableBool
+	FAASColdstart   nullable.Bool
 	FAASID          string
 	FAASName        string
 	FAASVersion     string