--- a/aggregators/nullable/doc.go
+++ b/aggregators/nullable/doc.go
@@ -1,0 +1,6 @@
+// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+// or more contributor license agreements. Licensed under the Elastic License 2.0;
+// you may not use this file except in compliance with the Elastic License 2.0.
+
+// Package nullable contains nullable types.
+package nullable